Progress(DiarySection)

+++ Code/DataControl/AccountControl.cs
@@ -17,7 +17,7 @@ namespace CaloRead
     public static class AccountControl
     {
 
-        public static bool AuthenticateLogin(string uname, string pword, ref double age, ref double weight, ref double height, ref string gender, ref double goal)
+        public static bool AuthenticateLogin(string uname, string pword, ref int age, ref float weight, ref float height, ref string gender, ref float goal)
         {
             HttpWebRequest request = (HttpWebRequest)WebRequest.Create($"http://192.168.1.6/caloread/login.php?uname={uname}&pword={pword} ");
             HttpWebResponse response = (HttpWebResponse)request.GetResponse();
@@ -37,11 +37,11 @@ namespace CaloRead
 
                 var user = root[0];
 
-                age = double.Parse(user.GetProperty("age").ToString());
-                weight = double.Parse(user.GetProperty("weight").ToString());
-                height = double.Parse(user.GetProperty("height").ToString());
+                age = int.Parse(user.GetProperty("age").ToString());
+                weight = float.Parse(user.GetProperty("weight").ToString());
+                height = float.Parse(user.GetProperty("height").ToString());
                 gender = user.GetProperty("gender").ToString();
-                goal = double.Parse(user.GetProperty("calorie_goal").ToString());
+                goal = float.Parse(user.GetProperty("calorie_goal").ToString());
 
 
                 return true;
@@ -52,7 +52,7 @@ namespace CaloRead
             }
         }
 
-        public static bool Register(string uname, string pword, double age, double weight, double height, string gender, double goal)
+        public static bool Register(string uname, string pword, int age, float weight, float height, string gender, float goal)
         {
             HttpWebRequest request = (HttpWebRequest)WebRequest.Create($"http://192.168.1.6/caloread/register.php?uname={uname}&pword={pword}&weight={weight}&height={height}&age={age}&gender={gender}&goal={goal}");
             HttpWebResponse response = (HttpWebResponse)request.GetResponse();
